+++ Assets/Scriptable Objects/BirdSmallPink.asset
@@ -12,7 +12,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 2f8ebddb3da3adc42bd9cb35220c9760, type: 3}
   m_Name: BirdSmallPink
   m_EditorClassIdentifier: 
-  spriteRenderer: {fileID: -6336752053224900323, guid: 11fa604a619fc47459a93991335e128a, type: 3}
+  sprite: {fileID: -6336752053224900323, guid: 11fa604a619fc47459a93991335e128a, type: 3}
   runtimeAnimatorController: {fileID: 9100000, guid: f4383a30e34f820489fe87ced2165103, type: 2}
   radius: 1.5
   mass: 5
