--- Assets/Scripts/BirdSelectingManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: bd96d7bb5ac5ff64b879d4b9975afab3
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
