+++ Assets/Scripts/CosmeticsWindow.cs
@@ -10,10 +10,14 @@ public class CosmeticsWindow : MonoBehaviour
     [SerializeField] private Transform readyButton;
     [SerializeField] private Transform leftButton;
     [SerializeField] private Transform rightButton;
+    [SerializeField] private BirdDatabaseSO birdDatabase;
+    [SerializeField] private Bird bird;
     // Start is called before the first frame update
+    private int selectedOption;
 
     private void Awake()
     {
+        selectedOption = 0;
         readyButton.GetComponent<Button_UI>().ClickFunc = () => {
             Hide();
             GameHandler.state = GameHandler.State.WaitingToStart;
@@ -21,12 +25,12 @@ public class CosmeticsWindow : MonoBehaviour
 
         leftButton.GetComponent<Button_UI>().ClickFunc = () =>
         {
-
+            BackOption();
         };
 
         rightButton.GetComponent<Button_UI>().ClickFunc = () =>
         {
-
+            NextOption();
         };
 
     }
@@ -36,9 +40,32 @@ public class CosmeticsWindow : MonoBehaviour
         gameObject.SetActive(false);
     }
 
-    // Update is called once per frame
-    void Update()
+    public void NextOption()
     {
-        
+        selectedOption++;
+        if (selectedOption >= birdDatabase.BirdCount)
+        {
+            selectedOption = 0;
+        }
+
+        UpdateBird(selectedOption);
     }
+
+    public void BackOption()
+    {
+        selectedOption--;
+        if (selectedOption < 0)
+        {
+            selectedOption = birdDatabase.BirdCount - 1;
+        }
+
+        UpdateBird(selectedOption);
+    }
+
+    private void UpdateBird(int selectedOption)
+    {
+        BirdSO birdSO = birdDatabase.GetBird(selectedOption);
+        bird.SetBirdSO(birdSO);
+    }
+
 }
